Correctly use pillar/grains/conf for SI
Previously we were only checking config (__opts__ if it was provided).
However, we don't just want to check the toplevel opts, but we also want
to check grains and pillar info.

As documented, the correct order is:

- env vars
- minion config
- minion grains
- minion pillar

This does that for get_service_instance, though other functions probably
don't quite have that same thing going for them.