Conf fixes (#307)
* add config layout ADR

* Add tests for utils.connect.get_service_instance

* Fix utils.connect.get_service_instance

Previously when using esxi_host or profiles, getting config values would
be problematic - either we'd use the wrong credentials or get
credentials from the wrong place. This has fixed the behavior to be
in-line with what we have documented:

- when connecting to vCenter, environment variables take precedence for
  host/user/password
- when connecting to ESXi, use the config values.
- if profiles are used, get ESXi credentials from the profile, rather
  than the top-level saltext.vmware config.

* Add esxi_host bug #306 changelog

* Add changelog, fix #256

* Fix esxi credentials

Previously we had a method that returned a user/pass, this new one
returns a dictionary. If we want the right values we've gotta get them
out. There also wasn't any test to correctly manage that, so...

* Convert missed service_instance calls

* Correctly use pillar/grains/conf for SI

Previously we were only checking config (__opts__ if it was provided).
However, we don't just want to check the toplevel opts, but we also want
to check grains and pillar info.

As documented, the correct order is:

- env vars
- minion config
- minion grains
- minion pillar

This does that for get_service_instance, though other functions probably
don't quite have that same thing going for them.