Change to more appropriate returns
Having old/new above the hosts was not correct - this state applies to
esxi hosts, so each host should be at the toplevel with an old/new key
containing the changes.

Also the existing output for changed values was unnecessarily verbose,
not too mention cumulatively comprehensive.

Tests needed some slight updates to be in-line with changes.